+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/data/ldap/LdapRepositoriesAutoConfigurationTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2017 the original author or authors.
+ * Copyright 2012-2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -73,7 +73,7 @@ public class LdapRepositoriesAutoConfigurationTests {
 				.applyTo(this.context);
 		this.context.register(configurationClasses);
 		this.context.register(LdapAutoConfiguration.class,
-				LdapDataAutoConfiguration.class, LdapRepositoriesAutoConfiguration.class,
+				LdapRepositoriesAutoConfiguration.class,
 				PropertyPlaceholderAutoConfiguration.class);
 		this.context.refresh();
 	}
