+++ spring-boot-project/spring-boot-test-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -15,7 +15,6 @@ org.springframework.boot.autoconfigure.transaction.TransactionAutoConfiguration
 
 # AutoConfigureDataLdap auto-configuration imports
 org.springframework.boot.test.autoconfigure.data.ldap.AutoConfigureDataLdap=\
-org.springframework.boot.autoconfigure.data.ldap.LdapDataAutoConfiguration,\
 org.springframework.boot.autoconfigure.data.ldap.LdapRepositoriesAutoConfiguration,\
 org.springframework.boot.autoconfigure.ldap.LdapAutoConfiguration,\
 org.springframework.boot.autoconfigure.ldap.embedded.EmbeddedLdapAutoConfiguration
